+++ src/types.ts
@@ -38,6 +38,7 @@ type BuncheeRollupConfig = Partial<Omit<RollupOptions, 'input' | 'output'>> & {
   exportName?: string
   input: InputOptions
   output: OutputOptions[]
+  dtsOnly: boolean
 }
 
 type CliArgs = CommonConfig & {
