+++ tsconfig.json
@@ -11,9 +11,9 @@
     "module": "CommonJS",
     "lib": ["ES6"],
     "removeComments": true,
-    // "strict": true,
-    // "noUnusedLocals": true,
-    // "noImplicitReturns": true,
+    "strict": true,
+    "noUnusedLocals": true,
+    "noImplicitReturns": true,
     "moduleResolution": "node",
     "resolveJsonModule": true,
     "esModuleInterop": true,
